* develop-fabernovel:
  Call api callback even if no read marker is sent
  Fix NPE when upload stats are null
  Change comparator used for read receipts
  Fix released session error when closing mxCrypto
  Add hook for room join method
  Fix NPE on MXSession#mEventsThread
  Add hook to add name disambiguator from host app
  Support excludeSelf when fetching read receipts if eventId is null
  Increase megolm rotation period
  Add API to check if forward pagination is possible
  Add listener to be notified when room is joined
  Add missing prev_sender in unsigned data

+++ matrix-sdk/src/main/java/org/matrix/androidsdk/crypto/algorithms/megolm/MXOutboundSessionInfo.java
@@ -47,7 +47,7 @@ public class MXOutboundSessionInfo {
         mUseCount = 0;
     }
 
-    public boolean needsRotation(int rotationPeriodMsgs, int rotationPeriodMs) {
+    public boolean needsRotation(int rotationPeriodMsgs, long rotationPeriodMs) {
         boolean needsRotation = false;
         long sessionLifetime = System.currentTimeMillis() - mCreationTime;
